mm: export symbol mm_access
Some devices can access process address spaces directly. When creating
such bond, to check that a process controlling the device is allowed to
access the target address space, the device driver uses mm_access(). Since
the drivers (in this case VFIO) can be built as a module, export the
mm_access symbol.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>