vfio/spimdev: Add documents for WarpDrive framework
WarpDrive is a common user space accelerator framework from Hisilicon.
The main component in Kernel is called spimdev, Share Parent IOMMU
Mediated Device. It exposes the hardware capabilities to the user space
via vfio-mdev. The user process can obtain a "queue" by open the device
and direct access and do DMA operation on the hardware via VFIO
interface.

The WarpDrive is intended to be used with Jean Philippe Brucker's SVA
patchset (it is still in RFC stage) to support multi-process. But it is
not a must. Without the patchset, WarpDrive can still work for one
process for every hardware device.

This patch add detail documents for the framework.

Signed-off-by: Kenneth Lee <liguozhu@hisilicon.com>