vfio: add spimdev framework
SPIMDEV is "Share Parent IOMMU Mdev". It is a vfio-mdev. But it shares
its parent's IOMMU and hardware resource is not allocated to the created mdev
but to the file when it is openned.

Currently only the vfio type1 driver is updated to let it be aware of
the existence of spimdev.

Signed-off-by: Kenneth Lee <liguozhu@hisilicon.com>
Signed-off-by: Zaibo Xu <xuzaibo@huawei.com>