crypto: Add Hisilicon Zip driver
The Hisilicon ZIP accelerator implements zlib and gzip algorithm support
for the software. It use Hisilicon QM as the interface to the CPU, so it
is shown up as a PCIE device to the CPU and communicate with he user
with queue communication.

This commit provides PCIE driver to the accelerator and register it to
the crypto subsystem.

Signed-off-by: Kenneth Lee <liguozhu@hisilicon.com>
Signed-off-by: Zhou Wang <wangzhou1@hisilicon.com>
Signed-off-by: Hao Fang <fanghao11@huawei.com>