acce4u: add accelerator-for-user framework
Acce4u is alternative solution of WarpDrive to spimdev. It dose not
rely on VFIO. It creates a new chrdev for the host device. When the
chrdev is openned, the host device's resource is allocated to it and dma
operation is conducted directly to the default domain of the host
device.