mm: export symbol find_get_task_by_vpid
Userspace drivers implemented with VFIO might want to bind sub-processes
to their devices. In a VFIO ioctl, they provide a pid that is used to find
a task and its mm. Since VFIO can be built as a module, export the
find_get_task_by_vpid symbol.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>

---
Here I didn't add a comment because neighbouring functions that get
exported don't have a comment either, and the name seems fairly clear.