uacce: support multiple Shared Virtual Address Space
SVAS, Shared Virtual Address Space, is a (maybe some in the future)
segment within a process.

It is created when the process mmap to a uacce queue from non-mmio
space.  When it is created, the virtual address space is valid to the
process and the uacce queue at the same time.

Other queue can be attached to the same address space to share the same
space.

The process can unmap the mapped address, but this will be removed the
SVAS.  The SVAS is removed until all queues that shared the SVAS are
closed.

The data structure is:

	q---->svas---->all_share_qs

The q->svas is protected by q->mutex. The svas->all_share_qs is
protected by svas->mutex. So it should be careful not to hold both mutex
at the same time, or you should only hold one queue mutex first and then
the svas.

If the device support SVA, SVAS is not necessary. The whole process is
shared with the device.

If the device support only IOMMU with pasid but without device fault
handling, SVA is supported

If the device support IOMMU without pasid and device fault, SVAS is
still supported. But more than one process to use the same IOMMU
instance may fail due to address confliction.

If the device do not support IOMMU, SVAS is still supported. But the
driver and the hardware should be able to translate the VA to valid
physical address by their self.

Signed-off-by: Kenneth Lee <liguozhu@hisilicon.com>